[준일] 리뷰Vo 수정 - int - > long

+++ src/main/java/hotelproject/repositories/vo/ReviewVo.java
@@ -8,7 +8,7 @@ import java.util.Date;
 @Getter
 @Setter
 public class ReviewVo {
-    private int userid; // 회원 ID
+    private Long userid; // 회원 ID
     private int hotelName;	// 호텔 이름
     private String reviewText; // 리뷰 내용
     private int rating; // 별점 (1~5)
